requirements.txt
fastapi>=0.75.0
uvicorn[standard]
redis-om>=0.0.20
python-decouple